Fix + Refactor Skills section
