+++ assets/stylesheets/sass/components/browser.sass
@@ -236,8 +236,6 @@ $header-color: rgb(56, 58, 55)
           fill: #eb3c00
 
       #sql
-        // color: #df6c20
-
         svg
           ellipse, path
             fill: none
