+++ index.html
@@ -2359,7 +2359,6 @@ Refer to https://opensource.org/licenses/MIT for a copy.
     </div> -->
   
     <script id="rendered-js" src="/dist/utilities/particles.js"></script>
-    <script src="/dist/components/skill.js" type="module"></script> 
     <script src="/dist/main.js" type="module" defer></script>
 
     <!-- TODO: Uncomment once "Mail" app is fixed -->
